+++ index.js
@@ -1,31 +1,32 @@
 'use strict'
 
-module.exports = {
-  fields: [],
-  codes: [],
-  messages: [],
-  field_feedback: {},
-  count: () => {
-    // How many errors are there?
-    // return this.errors.codes.length
-    return undefined
-  },
-  exists: () => {
-    // Are there any errors?
-    // if (this.errors_count > 0) {
-    //   return true
-    // } else {
-    //   return false
-    // }
-    return undefined
-  },
-  has_field: (field) => {
-    // Does it have a specific field?
-    // if (this.errors.fields.includes(field)) {
-    //   return true
-    // } else {
-    //   return false
-    // }
-    return undefined
+class ModelErrors {
+  constructor() {
+    this.fields = []
+    this.codes = []
+    this.messages = []
+    this.field_feedback = {}
+  }
+
+  get count() {
+    return this.codes.length
+  }
+
+  get exists() {
+    if (this.count > 0) {
+      return true
+    } else {
+      return false
+    }
+  }
+
+  has_field(field) {
+    if (this.fields.includes(field)) {
+      return true
+    } else {
+      return false
+    }
   }
 }
+
+module.exports = ModelErrors